-Documented "SplitScreenManager" and moved it to "utility" package
